create delete post api

+++ api/controllers/post.controller.js
@@ -14,7 +14,16 @@ export const getPost = async (req, res) => {
     const id = req.params.id;
     try {
         const post = await prisma.post.findUnique({
-            where: { id }
+            where: { id },
+            include: {
+                postDetail: true,
+                user: {
+                    select: {
+                        username: true,
+                        avatar: true
+                    }
+                }
+            }
         });
         res.status(200).json(post)
     } catch (err) {
@@ -58,17 +67,31 @@ export const deletePost = async (req, res) => {
 
     try {
         const post = await prisma.post.findUnique({
-            where: { id }
-        })
-        if (post.userId !== tokenUserId) {
-            res.status(403).json({ message: "Not Authorize!" })
+            where: { id },
+            include: { postDetail: true }
+        });
+
+        if (!post) {
+            return res.status(404).json({ message: "Post not found" });
         }
+
+        if (post.userId !== tokenUserId) {
+            return res.status(403).json({ message: "Not Authorized!" });
+        }
+
+        if (post.postDetail) {
+            await prisma.postDetail.delete({
+                where: { id: post.postDetail.id }
+            });
+        }
+
         await prisma.post.delete({
-            where: { id }
-        })
-        res.status(200).json({message: "Post deleted!"})
+            where: { id },
+        });
+
+        res.status(200).json({ message: "Post deleted" });
     } catch (err) {
-        console.log(err);
-        res.status(500).json({ message: "Failed to delete post!" })
+        console.error(err);
+        res.status(500).json({ message: "Failed to delete post" });
     }
-}
+};
